requirements.txt
autopep8==1.5.4
multidict==5.0.0
numpy==1.18.2
pandas==1.1.1
pylint==2.6.0
pytest==6.0.1
click==7.1.2
python-dateutil==2.8.1
python-dotenv==0.14.0
wntr==0.3.0
keras==2.4.3
tensorflow==1.15.4
tensorboard==1.15.0
tensorboardX==2.2
matplotlib==3.3.3
networkx==2.5
geopandas==0.8.1
osmnx==0.15.0
Rtree==0.9.4
setuptools==49.6.0
pmdarima==1.8.0
mlflow==1.13.1
dvc==1.11.10
bcrypt==3.2.0
paramiko==2.7.2
modAL==0.4.1
plotly==4.14.3
knack==0.7.2
dtreeviz==1.2
gym==0.17.2
stable-baselines==2.10.0
tables==3.6.1
ray==1.5.1
tree==0.2.4
dm-tree==0.1.6
pysftp==0.2.9
pytest-mock==3.6.1
sphinx==4.3.0
sphinx-rtd-theme==1.0.0
# mpi4py==3.0.3
PyYAML
scipy
python-dotenv
mpi4py